Move bookshelf.js to models dir (#75)
