+++ models/tournament.js
@@ -1,4 +1,4 @@
-var bookshelf = require("../bookshelf"),
+var bookshelf = require("./bookshelf"),
 	checkit = require('checkit');
 
 var Tournament = bookshelf.Model.extend({
