+++ services/games.js
@@ -1,6 +1,6 @@
 
 var app = require("../services/app");
-var bookshelf = require("../bookshelf");
+var bookshelf = require("../models/bookshelf");
 var Stat = require("./stats");
 var Team = require("./teams");
 var adminOnly = require( "../adminOnly" );
